refactor
Added additional header files for future convenience
